Uploaded 9 gene lists for feature selection
Gene lists have different thrresholds for filtering. First filtering factor is by the fraction of cells in which the gene is unexpressed. The higher the fraction, the more likely this gene is to be a "signature gene" of a particular cell type. highfraction : > 0.8 proprortion, mediumfraction : > 0.2, lowfraction : >0. The second filtering factor is by variance. highvariance : top 1000, mediumvariance : top 3000, lowvariance : top 5000. The genelists and file to produce them are provided.